# openapi.yaml
openapi: 3.1.0
info:
  title: KORION KORI Backend API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/onchain/networks/{network}/wallets/{address}/balance:
    get:
      summary: Lookup on-chain TRX and KORI balance for any address on mainnet or testnet
      parameters:
        - in: path
          name: network
          required: true
          schema:
            type: string
            enum: [mainnet, testnet]
        - in: path
          name: address
          required: true
          schema:
            type: string
            pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
      responses:
        '200':
          description: On-chain balance lookup response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnchainBalanceResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /api/onchain/networks/{network}/transfers:
    post:
      summary: Send KORI directly from hot wallet on the selected network
      parameters:
        - in: path
          name: network
          required: true
          schema:
            type: string
            enum: [mainnet, testnet]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnchainTransferRequest'
      responses:
        '201':
          description: On-chain hot wallet transfer accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnchainTransferResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    WalletMonitoringSnapshot:
      type: object
      properties:
        address:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        tokenSymbol:
          type: string
        tokenContractAddress:
          type: string
          nullable: true
        tokenBalance:
          type: string
          nullable: true
        tokenRawBalance:
          type: string
          nullable: true
        tokenDecimals:
          type: integer
          nullable: true
        trxBalance:
          type: string
          nullable: true
        trxRawBalance:
          type: string
          nullable: true
        fetchedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [ok, error]
        error:
          type: string
          nullable: true
    OnchainTransferRequest:
      type: object
      required: [toAddress, amount]
      properties:
        toAddress:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        amount:
          type: number
          minimum: 0
    OnchainBalanceResponse:
      type: object
      properties:
        network:
          type: string
          enum: [mainnet, testnet]
        tronApiUrl:
          type: string
        tokenContractAddress:
          type: string
        wallet:
          $ref: '#/components/schemas/WalletMonitoringSnapshot'
    OnchainTransferResponse:
      type: object
      properties:
        network:
          type: string
          enum: [mainnet, testnet]
        tronApiUrl:
          type: string
        tokenContractAddress:
          type: string
        hotWalletAddress:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        gatewayMode:
          type: string
        toAddress:
          type: string
          pattern: '^T[1-9A-HJ-NP-Za-km-z]{33}$'
        amount:
          type: string
        txHash:
          type: string
        sentAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true
...